move ensure exec tests
